【Feature/admindashboreds】fixed some conflicts
