+++ app/Http/Controllers/AdminController.php
@@ -11,7 +11,6 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Response;
-use App\Models\User;
 
 class AdminController extends Controller
 {
@@ -21,9 +20,8 @@ class AdminController extends Controller
     private $review;
     private $reservation;
 
-    private $user;
 
-    public function __construct(Profile $profile, User $user, Restaurant $restaurant, Review $review, Reservation $reservation, User $user,){
+    public function __construct(Profile $profile, User $user, Restaurant $restaurant, Review $review, Reservation $reservation){
         $this->profile = $profile;
         $this->user = $user;
         $this->restaurant = $restaurant;
@@ -126,82 +124,4 @@ class AdminController extends Controller
         return back();
     }
 
-    public function userChart(){
-        $users = User::selectRaw('MONTH(created_at) as month, COUNT(*) as count')
-                    ->whereYear('created_at', date('Y'))
-                    ->groupBy('month')
-                    ->oderBy('month')
-                    ->get();
-
-        $labels = [];
-        $data = [];
-
-        for($i=1; $i < 12; $i++) {
-            $month = date('F',mktime(0,0,0,$i,1));
-            $count =0;
-
-            foreach($users as $user){
-                if($user->month == $i){
-                    $count =$user->count;
-                    break;
-                }
-            }
-
-            array_push($labels,$month);
-            array_push($data,$count);
-        }
-
-        $datasets = [
-            [
-                'label' => 'Users',
-                'data' => $data,
-
-            ]
-            ];
-            return view('admin.dashboard',compact('datasets','labels'));
-    }
-
-    public function userChartApi(Request $request){
-        $user_type = $request->user_type;
-        
-        $users = [];
-        $labels = [];
-        $data = [];
-
-        switch ($user_type){
-            case 'user':
-                $users = DB::table('profiles')->where('usertype_id', '=', 2)->get();
-                break;
-            
-        }
-
-        for($i=1; $i < 13; $i++) {
-            $month = date('M',mktime(0,0,0,$i,1));
-
-
-            array_push($labels,$month);
-
-        }
-
-        for($i=1; $i < 13; $i++){
-            $userCount = 0;
-
-            for($j = 0; $j < count($users); $j++){
-                $userMonthCreated = (int)explode('-', $users[$j]->created_at)[1];
-                if($userMonthCreated == $i){
-                    $userCount++;
-                }
-            }
-            array_push($data,$userCount);
-        }
-
-        $datasets = [
-            [
-                'label' => 'Users',
-                'data' => $data,
-                'backgroundColor' => "#CAC2C7"
-            ]
-        ];
-        return Response::json(['success'=>true, 'labels'=>$labels, 'datasets'=>$datasets]);
-    }
 }
